@@ -31,7 +31,7 @@ public abstract class IndexedBlockStatePredicate implements Predicate<BlockState
         if (this.index > 31 || this.index >= BlockStateFlags.NUM_FLAGS) {
             throw new IndexOutOfBoundsException();
         }
-        this.mask = 1 << index;
+        this.mask = 1 << this.index;
 
         //initialization is run on one thread with synchronization afterwards, so escaping this here is fine
         ALL_FLAGS[this.index] = this;
@@ -42,10 +42,10 @@ public abstract class IndexedBlockStatePredicate implements Predicate<BlockState
     }
 
     public int getIndex() {
-        return index;
+        return this.index;
     }
 
     public int getMask() {
-        return mask;
+        return this.mask;
     }
 }
