@@ -1,7 +1,6 @@
 package me.jellysquid.mods.lithium.mixin.block.moving_block_shapes;
 
 import me.jellysquid.mods.lithium.common.shapes.OffsetVoxelShapeCache;
-import me.jellysquid.mods.lithium.common.util.tuples.FinalObject;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.shape.VoxelShape;
 import org.spongepowered.asm.mixin.Mixin;
@@ -9,32 +8,30 @@ import org.spongepowered.asm.mixin.Mixin;
 
 @Mixin(VoxelShape.class)
 public class VoxelShapeMixin implements OffsetVoxelShapeCache {
-    private FinalObject<VoxelShape>[] offsetAndSimplified;
+    private volatile VoxelShape[] offsetAndSimplified;
 
     public void setShape(float offset, Direction direction, VoxelShape offsetShape) {
         if (offsetShape == null) {
             throw new IllegalArgumentException("offsetShape must not be null!");
         }
         int index = getIndexForOffsetSimplifiedShapes(offset, direction);
-        FinalObject<VoxelShape>[] offsetAndSimplified = this.offsetAndSimplified;
-        if (offsetAndSimplified == null) {
-            //noinspection unchecked
-            this.offsetAndSimplified = (offsetAndSimplified = new FinalObject[1 + (2 * 6)]);
+        VoxelShape[] offsetAndSimplifiedShapes = this.offsetAndSimplified;
+        if (offsetAndSimplifiedShapes == null) {
+            offsetAndSimplifiedShapes = new VoxelShape[1 + 2 * 6];
+        } else {
+            offsetAndSimplifiedShapes = offsetAndSimplifiedShapes.clone();
         }
-        //using FinalObject as it stores the value in a final field, which guarantees safe publication
-        offsetAndSimplified[index] = new FinalObject<>(offsetShape);
+        offsetAndSimplifiedShapes[index] = offsetShape;
+        this.offsetAndSimplified = offsetAndSimplifiedShapes;
     }
 
     public VoxelShape getOffsetSimplifiedShape(float offset, Direction direction) {
-        FinalObject<VoxelShape>[] offsetAndSimplified = this.offsetAndSimplified;
+        VoxelShape[] offsetAndSimplified = this.offsetAndSimplified;
         if (offsetAndSimplified == null) {
             return null;
         }
         int index = getIndexForOffsetSimplifiedShapes(offset, direction);
-        //usage of FinalObject guarantees that we are seeing a fully initialized VoxelShape here, even when it was created on a different thread
-        FinalObject<VoxelShape> wrappedShape = offsetAndSimplified[index];
-        //noinspection FinalObjectAssignedToNull,FinalObjectGetWithoutIsPresent
-        return wrappedShape == null ? null : wrappedShape.getValue();
+        return offsetAndSimplified[index];
     }
 
     private static int getIndexForOffsetSimplifiedShapes(float offset, Direction direction) {
