@@ -9,6 +9,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
+import java.util.function.Predicate;
 import java.util.stream.Stream;
 
 @Mixin(Entity.class)
@@ -38,9 +39,23 @@ public abstract class EntityMixin {
     private Stream<BlockState> skipFireTestIfResultDoesNotMatter(World world, Box box) {
         // Skip scanning the blocks around the entity touches by returning an empty stream when the result does not matter
         if ((this.fireTicks > 0 || this.fireTicks == -this.getBurningDuration()) && (!this.wasOnFire || !this.inPowderSnow && !this.isWet())) {
-            return Stream.empty();
+            return null;
         }
 
         return world.getStatesInBoxIfLoaded(box);
+    }
+
+    @Redirect(
+            method = "move(Lnet/minecraft/entity/MovementType;Lnet/minecraft/util/math/Vec3d;)V",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Ljava/util/stream/Stream;noneMatch(Ljava/util/function/Predicate;)Z"
+            )
+    )
+    private boolean skipNullStream(Stream<BlockState> stream, Predicate<BlockState> predicate) {
+        if (stream == null) {
+            return true;
+        }
+        return stream.noneMatch(predicate);
     }
 }
