@@ -1,5 +1,7 @@
 package me.jellysquid.mods.lithium.mixin.gen.cached_generator_settings;
 
+import net.minecraft.util.math.noise.DoublePerlinNoiseSampler;
+import net.minecraft.util.registry.Registry;
 import net.minecraft.world.biome.source.BiomeSource;
 import net.minecraft.world.gen.chunk.ChunkGeneratorSettings;
 import net.minecraft.world.gen.chunk.NoiseChunkGenerator;
@@ -12,6 +14,9 @@ import java.util.function.Supplier;
 
 @Mixin(NoiseChunkGenerator.class)
 public class NoiseChunkGeneratorMixin {
+    @Shadow
+    @Final
+    protected Supplier<ChunkGeneratorSettings> settings;
     private int cachedSeaLevel;
 
     /**
@@ -30,14 +35,14 @@ public class NoiseChunkGeneratorMixin {
      * Initialize the cache early in the ctor to avoid potential future problems with uninitialized usages
      */
     @Inject(
-            method = "<init>(Lnet/minecraft/world/biome/source/BiomeSource;Lnet/minecraft/world/biome/source/BiomeSource;JLjava/util/function/Supplier;)V",
+            method = "<init>(Lnet/minecraft/util/registry/Registry;Lnet/minecraft/world/biome/source/BiomeSource;Lnet/minecraft/world/biome/source/BiomeSource;JLjava/util/function/Supplier;)V",
             at = @At(
                     value = "INVOKE",
                     target = "Lnet/minecraft/world/gen/chunk/ChunkGeneratorSettings;getGenerationShapeConfig()Lnet/minecraft/world/gen/chunk/GenerationShapeConfig;",
                     shift = At.Shift.BEFORE
             )
     )
-    private void hookConstructor(BiomeSource populationSource, BiomeSource biomeSource, long seed, Supplier<ChunkGeneratorSettings> settings, CallbackInfo ci) {
-        this.cachedSeaLevel = settings.get().getSeaLevel();
+    private void hookConstructor(Registry<DoublePerlinNoiseSampler.NoiseParameters> noiseRegistry, BiomeSource biomeSource, BiomeSource biomeSource2, long seed, Supplier<ChunkGeneratorSettings> supplier, CallbackInfo ci) {
+        this.cachedSeaLevel = this.settings.get().getSeaLevel();
     }
 }
