@@ -2,26 +2,29 @@ package me.jellysquid.mods.lithium.mixin.shapes.lazy_shape_context;
 
 import net.minecraft.block.EntityShapeContext;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.fluid.FlowableFluid;
 import net.minecraft.fluid.Fluid;
 import net.minecraft.fluid.FluidState;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import org.spongepowered.asm.mixin.*;
-import org.spongepowered.asm.mixin.injection.*;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Constant;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyConstant;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import java.util.Optional;
 import java.util.function.Predicate;
 
 @Mixin(EntityShapeContext.class)
 public class EntityShapeContextMixin {
-    @SuppressWarnings("OptionalUsedAsFieldOrParameterType")
-    @Shadow
-    @Final
-    private Optional<Entity> entity;
-
     @Mutable
     @Shadow
     @Final
@@ -32,10 +35,10 @@ public class EntityShapeContextMixin {
     @Final
     private Predicate<Fluid> walkOnFluidPredicate;
 
-    @Mutable
     @Shadow
     @Final
-    private ItemStack boots;
+    @Nullable
+    private Entity entity;
 
     /**
      * Mixin the instanceof to always return false to avoid the expensive inventory access.
@@ -57,47 +60,40 @@ public class EntityShapeContextMixin {
         return false;
     }
 
-    @ModifyConstant(
-            method = "<init>(Lnet/minecraft/entity/Entity;)V",
-            constant = @Constant(classValue = LivingEntity.class, ordinal = 4)
-    )
-    private static boolean redirectInstanceOf3(Object obj, Class<?> clazz) {
-        return false;
-    }
-
     @Inject(
             method = "<init>(Lnet/minecraft/entity/Entity;)V",
             at = @At(
                     value = "INVOKE",
-                    target = "Lnet/minecraft/block/EntityShapeContext;<init>(ZDLnet/minecraft/item/ItemStack;Lnet/minecraft/item/ItemStack;Ljava/util/function/Predicate;Ljava/util/Optional;)V",
+                    target = "Lnet/minecraft/block/EntityShapeContext;<init>(ZDLnet/minecraft/item/ItemStack;Ljava/util/function/Predicate;Lnet/minecraft/entity/Entity;)V",
                     shift = At.Shift.AFTER
             )
     )
     private void initFields(Entity entity, CallbackInfo ci) {
         this.heldItem = null;
         this.walkOnFluidPredicate = null;
-        this.boots = null;
     }
 
-    /**
-     * @author 2No2Name
-     * @reason allow skipping unused initialization
-     */
-    @Overwrite
-    public boolean isHolding(Item item) {
+    @Inject(
+            method = "isHolding(Lnet/minecraft/item/Item;)Z",
+            at = @At("HEAD")
+    )
+    public void isHolding(Item item, CallbackInfoReturnable<Boolean> cir) {
         if (this.heldItem == null) {
-            this.heldItem = this.entity.isPresent() && this.entity.get() instanceof LivingEntity ? ((LivingEntity)this.entity.get()).getMainHandStack() : ItemStack.EMPTY;
+            this.heldItem = this.entity instanceof LivingEntity ? ((LivingEntity) this.entity).getMainHandStack() : ItemStack.EMPTY;
         }
-        return this.heldItem.isOf(item);
     }
 
-
-    /**
-     * @author 2No2Name
-     * @reason allow skipping unused lambda allocation
-     */
-    @Overwrite
-    public boolean canWalkOnFluid(FluidState aboveState, FlowableFluid fluid) {
-        return this.entity.isPresent() && this.entity.get() instanceof LivingEntity && ((LivingEntity) this.entity.get()).canWalkOnFluid(fluid) && !aboveState.getFluid().matchesType(fluid);
+    @Inject(
+            method = "canWalkOnFluid(Lnet/minecraft/fluid/FluidState;Lnet/minecraft/fluid/FlowableFluid;)Z",
+            at = @At("HEAD")
+    )
+    public void canWalkOnFluid(FluidState state, FlowableFluid fluid, CallbackInfoReturnable<Boolean> cir) {
+        if (this.walkOnFluidPredicate == null) {
+            if (this.entity instanceof LivingEntity livingEntity) {
+                this.walkOnFluidPredicate = livingEntity::canWalkOnFluid;
+            } else {
+                this.walkOnFluidPredicate = (liquid) -> false;
+            }
+        }
     }
 }
